move use-geolocation to hooks folder
