+++ packages/template-retail-react-app/app/components/store-locator/form.tsx
@@ -17,7 +17,7 @@ import {
 } from '@chakra-ui/react'
 import {useForm, Controller} from 'react-hook-form'
 import {useStoreLocator} from '@salesforce/retail-react-app/app/hooks/use-store-locator'
-import {useGeolocation} from '@salesforce/retail-react-app/app/components/store-locator/use-geo-location'
+import {useGeolocation} from '@salesforce/retail-react-app/app/hooks/use-geo-location'
 import type {StoreLocatorFormValues} from '@salesforce/retail-react-app/app/components/store-locator/types'
 
 export const StoreLocatorForm: React.FC = () => {
